used Clion to pick some lint. It had enough misfires that I am back to qtcreator.

+++ cppext/quadraticfilter.cpp
@@ -40,7 +40,7 @@ double QuadraticFilter::amplitude() const {
   return amplitude(Y[0]);
 }
 double QuadraticFilter::amplitude(int datum) const {
-  return ratio(double(S4 * datum - S2 * Y[2]),D4);
+  return ratio(S4 * datum - S2 * Y[2], D4);
 }
 
 int QuadraticFilter::ampEstimate() const {
@@ -54,8 +54,8 @@ void QuadraticFilter::recordInflection(Inflection &flect) const {
 
 void QuadraticFilter::init(const CenteredSlice &slice){
   //accumulators
-  Y[2] = 0.0;
-  Y[1] = 0.0;
+  Y[2] = 0;
+  Y[1] = 0;
   Y[0] = slice[0];
   for(int fi = 1; fi<=hw; ++fi) {
     int high = slice[fi];
